backend/requirements.txt
openai>=1.0.0
pinecone-client>=2.2.4
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
black>=23.11.0
supabase-py>=2.0.0
langchain>=0.1.0
requests>=2.31.0
telethon>=1.33.1
pytest>=7.4.3 